requirement.txt
customtkinter
tkinter
Pillow
opencv-python-headless
numpy
paddleocr
scikit-learn
joblib
sqlite3